enable cache abnnout TRN network buolder

+++ R/annotations/transcript_network/tfbs_motif_scanning.R
@@ -72,20 +72,25 @@ const tfbs_motif_scanning = function(app, context) {
                 let model_id = basename(model_dir);
                 let upstream_seq = file.path(source_dir, model_id, "upstream_locis.fasta");
                 let outfile = file.path(workdir, model_id, "tfbs_motifs.csv");
-                # make TFBS site scanning on the TSS upstream region sites
-                # search site against the reference motif search.
-                let motifs = GCModeller::scan_motifs(
-                    db = motifs_db, 
-                    seqs = upstream_seq, 
-                    workdir = dirname(outfile), 
-                    n_threads = n_threads,
-                    pval_cutoff = motif_pvalcut,
-                    scan_reverse = FALSE
-                );
 
-                message(`motif site scan success for ${model_id}, found ${nrow(motifs)} motif site matches!`);
+                if (!file.exists(outfile)) {
+                    # make TFBS site scanning on the TSS upstream region sites
+                    # search site against the reference motif search.
+                    let motifs = GCModeller::scan_motifs(
+                        db = motifs_db, 
+                        seqs = upstream_seq, 
+                        workdir = dirname(outfile), 
+                        n_threads = n_threads,
+                        pval_cutoff = motif_pvalcut,
+                        scan_reverse = FALSE
+                    );
 
-                write.csv(motifs, file = outfile, row.names = FALSE);
+                    message(
+                        `motif site scan success for ${model_id}, found ${nrow(motifs)} motif site matches!`
+                    );
+
+                    write.csv(motifs, file = outfile, row.names = FALSE);
+                }
             }
         } else {
             # set the TSS upstream region site fasta file
